==> 01_preprocess_fail.ipynb <==
### Read Csv

import pandas
raw = pandas.read_csv("./data_kaggle/train.csv");

raw

### Acquiring the dataset

Ideally the dataset should already be labelled, is large enough and contains posts about a variety of disasters.
My First approach to this was to use [the dataset provided over at Kaggle's competition, 'Natural Language Processing with Disaster Tweets'](https://www.kaggle.com/competitions/nlp-getting-started/overview), which claims to contain about 10k labelled tweets that may or may not be related to actual disasters.

Each row of data contains the following columns:
* "target" indicates whether the tweet is about an actual disaster happening
* "keyword" indicates the keyword related to the disaster (has missing values)
* "location" indicates location provided by the user with the tweet (has missing values)

From the description, it is clear that location does not hold much point since we want the trained model to work anywhere, not just in a certain location.

#### Taking a peek at the values

keywords = raw["keyword"].value_counts().index.to_list();
keywords.sort()
keywords

Looking at the keywords in the original dataset, it is clear that:
* Not all keywords are meaningful for predicting disasters. For example, "armageddon" is very unlikely to happen, and "battle" mostly contains gibberish about fictional "wars"
* The keywords may not necessarily indicate the kind of disaster; For example, 'survive', 'survivors' does not really mean anything
* A lot of words are just different forms of the same thing: "buildings burning", "buildings on fire", "bush fires", "fire", "fire truck", "forest fire", "forest fires", "hellfire", "ablaze" is just a different way of saying "fire"

* A lot of tweets contain URLs, which points to either- A: Twitter medias B: external websites

It is also clear that:
* A lot of data does not indicate "actual" disaster; id 6132 "The Prophet (peace be upon him) said 'Save yourself from Hellfire even if it is by giving half a date in charity.' is from Islamic Hadith, however is marked as a "disaster"
* While the dataset provider claims that the data is "hand-labelled", it is in fact not true: [There are discussions suggesting that some sort of heuristic is used to label the sentences](https://www.kaggle.com/competitions/nlp-getting-started/discussion/130458).

While it would be very convenient to use pre-labelled data, with the above points we can conclude that using this data would be largely impractical and will not help us get reliable results regarding classification of 'actual disasters'.

Therefore, unfortunately, it is decided that we would have to collect the data ourselves and label them.

### Collecting & labelling data manually

We collect data from DCInside, by searching for specific terms using the search page.
For each term we collect 20 pages worth of data (20 pages * 20 articles = 400 posts).
For each post, we collect:
* The gallery name
* The title
* Part of the post (text) content
* Post date

which we add label of:
* Search keyword
* Whether the post is about an actual disaster or not
  * This does not necessarily have to be about current disasters; for example, posts talking about past disasters
