==> 02_collection.ipynb <==
from dataclasses import dataclass
from typing import Literal, Optional
import numpy
import pandas

@dataclass
class Post:
    gallery: str
    title: str
    content: str
    date: numpy.datetime64
    keyword: str
    searchMode: Literal["accuracy", "latest"]
    disaster: Optional[bool]

from typing import Optional
import requests
from urllib.parse import quote
import bs4

@dataclass
class SearchResultPost:
    gallery: str
    title: str
    content: str
    date: numpy.datetime64

def search(query: str, page: int = 1, mode: Literal["latest", "accuracy"] = "accuracy") -> Optional[list[SearchResultPost]]:
    response = requests.get(f"https://search.dcinside.com/post/p/{page}/sort/{mode}/q/{quote(query, safe='.')}")
    if not (200 <= response.status_code < 299):
        print(response.text);
        return None
    
    list = []
    soup = bs4.BeautifulSoup(response.text)
    for element in soup.select(".sch_result_list>li"):
        gallery = element.select_one(".sub_txt").get_text()
        title = element.select_one(".tit_txt").get_text()
        content = element.select_one(".link_dsc_txt").get_text()
        date = pandas.to_datetime(element.select_one(".date_time").get_text())
        list.append(SearchResultPost(gallery, title, content, date))
    return list

from time import sleep

keywords = ["지진", "화재", "불남", "싱크홀", "근데"]

def scrapeKeyword(keyword: str, mode = Literal["accuracy", "latest"], pages: int = 5):
    labeledPosts = []
    for page in range(1, pages + 1):
        print(f"Keyword {keyword}, page {page}, mode {mode}")
        posts = search(keyword, page, mode)
        sleep(1)
        if posts == None:
            raise Exception(f"Failed to fetch search results for keyword {keyword}, page {page}")

        for post in posts:
            # print(f"{post.title} - {post.gallery} {post.date:%Y-%m-%d}\n{post.content}")
            labeledPosts.append(Post(gallery=post.gallery, title=post.title, content=post.content, date=post.date, keyword=keyword, disaster=None, searchMode=mode))
    return labeledPosts

posts = []

for keyword in keywords:
    posts += scrapeKeyword(keyword, mode="accuracy");
    posts += scrapeKeyword(keyword, mode="latest");

df = pandas.DataFrame(posts)

df

df.to_csv("./data_dcinside/raw_data.csv")
