==> 03_1_preprocess.ipynb <==
import pandas

df = pandas.read_csv("./data_dcinside/raw_data.csv", index_col=0)

def clean_data(df):
    # Drop duplicate rows in columns: 'title', 'content'
    df = df.drop_duplicates(subset=['title', 'content'])
    return df

df_clean = clean_data(df.copy())
df_clean.head()

len(df_clean)

df_clean.to_csv("./data_dcinside/data_deduped.csv")
