==> 03_2_label.ipynb <==
%sudo apt-get install g++ openjdk-8-jdk python3-dev python3-pip curl
%python3 -m pip install konlpy
%pip install mecab-python3
%bash <(curl -s https://raw.githubusercontent.com/konlpy/konlpy/master/scripts/mecab.sh)

from time import sleep
from konlpy.tag import Kkma, Okt, Mecab
tagger = Mecab()
# Look for NNG (보통 명사) '지진'
sampleText = "전쟁...기후 변화...지진 모든 것이 성경에 새겨진 종말의 징조다 그때가 오면 악마들이 온다 ........... 악마는 예전에도 지금도 미래에도 있을 것인데 우리가 지금 악마의 시대에 살고 있다"
if "지진/NNG" in tagger.pos(sampleText, join=True):
    print("yes")
else:
    print("no")

# 화재/NNG
# 싱크홀/NNP
tagger.pos("방지턱 없는 데 차들 ‘덜컹’…연희동 싱크홀 10분 전 상황(영상) 서울 서대문구 연희동 4차선 도로에서 싱크홀(땅꺼짐 현상)이 발생해 SUV 차량에 타고 있던 운전자 등 2명이 부상을 당한 가운데 사고가...")


from dataclasses import dataclass
from typing import Literal, Optional
import numpy
import pandas

@dataclass
class Post:
    gallery: str
    title: str
    content: str
    date: numpy.datetime64
    keyword: str
    searchMode: Literal["accuracy", "latest"]
    disaster: Optional[bool]

df = pandas.read_csv("./data_dcinside/data_deduped.csv", index_col=0, dtype={"content": str, "title": str})

df.dtypes

keywordToTag = {
    "지진": "지진/NNG",
    "화재": "화재/NNG",
    "싱크홀": "싱크홀/NNP"
}

def crudeLabel(row):
    tag = keywordToTag.get(row.keyword)
    joinedContent = str(row.title) + '\n' + str(row.content)
    if tag:
        if tag in tagger.pos(joinedContent, join=True):
            return True
        else:
            return False
    return None

labels = list(map(crudeLabel, df.itertuples()))
df["disaster"] = labels

df

df.to_csv("./data_dcinside/data_crude_label.csv")
