==> 03_3_transform.ipynb <==
import pandas

df = pandas.read_csv("./data_dcinside/data_hand_labelled.csv", index_col=0)

df

f = df["disaster"] == True

df.loc[f, "disasterKind"] = df.loc[f]["keyword"]

df["disasterKind"] = df["disasterKind"].replace({
    "지진": "earthquake",
    "화재": "fire",
    "불남": "fire",
    "싱크홀": "sinkhole",
}, )

df.loc[df["disasterKind"].isna(), "disasterKind"] = "none"

df

df

df["merged"] = df["title"].map(lambda x: str(x)) + "\n" + df["content"].map(lambda x: str(x))

transformed = df[["merged", "disasterKind"]]

transformed

transformed.to_csv("./data_dcinside/transformed.csv")

transformed["disasterKind"].unique()

transformed["disasterKind"].value_counts()
